Enhancements and Fixes for Runtime Errors

+++ src/Core/Blog.Application/CQRS/Commands/BlogPost/DeleteBlogPost/DeleteBlogPostHandler.cs
@@ -1,23 +1,27 @@
-﻿using Blog.Application.Repositories;
-using Blog.Application.Wrappers;
+﻿using Blog.Application.Exceptions;
+using Blog.Application.Repositories;
 using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using ValidationException = FluentValidation.ValidationException;
 
 namespace Blog.Application.CQRS.Commands.BlogPost.DeleteBlogPost
 {
     public class DeleteBlogPostHandler : IRequestHandler<DeleteBlogPostRequest, DeleteBlogPostResponse>
     {
         private readonly IBlogPostWriteRepository _blogPostWriteRepository;
+        private readonly IBlogPostReadRepository _blogPostReadRepository;
         private readonly ILogger<DeleteBlogPostHandler> _logger;
         private readonly IValidator<DeleteBlogPostRequest> _validator;
 
         public DeleteBlogPostHandler(
             IBlogPostWriteRepository blogPostWriteRepository,
+            IBlogPostReadRepository blogPostReadRepository,
             ILogger<DeleteBlogPostHandler> logger,
             IValidator<DeleteBlogPostRequest> validator)
         {
             _blogPostWriteRepository = blogPostWriteRepository;
+            _blogPostReadRepository = blogPostReadRepository;
             _logger = logger;
             _validator = validator;
         }
@@ -34,11 +38,16 @@ namespace Blog.Application.CQRS.Commands.BlogPost.DeleteBlogPost
                 throw new ValidationException(validationResult.Errors);
             }
 
-            await _blogPostWriteRepository.RemoveAsync(request.Id);
+            var blogPostToBeDeleted = await _blogPostReadRepository.GetByIdAsync(request.Id);
+
+            if (blogPostToBeDeleted == null)
+                throw new EntityNotFoundException(nameof(BlogPost), request.Id);
+
+            _blogPostWriteRepository.Remove(blogPostToBeDeleted);
 
             var response = new DeleteBlogPostResponse()
             {
-                Message = $"BlogPost with Id: {request.Id} has been deleted"
+                Value = $"BlogPost with Id: {request.Id} has been deleted"
             };
 
             return response;
